Move clean_json_response helper function and corresponding tests
