+++ cyber_query_ai/helpers.py
@@ -1,0 +1,25 @@
+"""Helper methods for the CyberQueryAI application."""
+
+import re
+
+
+def clean_json_response(response_text: str) -> str:
+    """Clean common JSON formatting issues from LLM responses."""
+    # Remove trailing commas in arrays and objects
+    response_text = re.sub(r",(\s*[}\]])", r"\1", response_text)
+
+    # Remove any markdown code blocks if present
+    response_text = re.sub(r"```json\s*", "", response_text)
+    response_text = re.sub(r"```\s*$", "", response_text)
+
+    # Fix common structural issues where explanation is inside commands array
+    # Pattern: ["command", "explanation": "text"] -> ["command"], "explanation": "text"
+    response_text = re.sub(
+        r'(\["[^"]*"\s*),\s*"(explanation?)":\s*"([^"]*)"(\s*\])',
+        r'\1], "\2": "\3"',
+        response_text,
+        flags=re.IGNORECASE,
+    )
+
+    # Strip whitespace
+    return response_text.strip()
